+++ src/pytorch_lightning/strategies/horovod.py
@@ -31,7 +31,6 @@ from pytorch_lightning.utilities.distributed import ReduceOp
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.imports import _HOROVOD_AVAILABLE
 from pytorch_lightning.utilities.rank_zero import rank_zero_only
-from pytorch_lightning.utilities.types import _LRScheduler
 
 if _HOROVOD_AVAILABLE:
     import horovod.torch as hvd
@@ -114,8 +113,8 @@ class HorovodStrategy(ParallelStrategy):
         lr_scheduler_configs = self.lr_scheduler_configs
         for config in lr_scheduler_configs:
             scheduler = config.scheduler
-            assert isinstance(scheduler, _LRScheduler)
-            scheduler.base_lrs = [lr * self.world_size for lr in scheduler.base_lrs]
+            if hasattr(scheduler, "base_lrs"):
+                scheduler.base_lrs = [lr * self.world_size for lr in scheduler.base_lrs]  # type: ignore[union-attr]
 
         assert self.lightning_module is not None
         # Horovod: broadcast parameters & optimizer state to ensure consistent initialization